+++ metadata.rb
@@ -13,7 +13,7 @@ end
 source_url 'https://github.com/djoos-cookbooks/newrelic' if respond_to?(:source_url)
 issues_url 'https://github.com/djoos-cookbooks/newrelic/issues' if respond_to?(:issues_url)
 
-depends 'python'
+depends 'poise-python'
 depends 'curl'
 depends 'apt'
 depends 'yum'