+++ providers/agent_python.rb
@@ -24,17 +24,19 @@ action :install do
 end
 
 action :remove do
-  python_pip 'newrelic' do
+  python_package 'newrelic' do
     virtualenv new_resource.virtualenv if new_resource.virtualenv
     version new_resource.version if new_resource.version
+    python new_resource.python if new_resource.python
     action :remove
   end
 end
 
 def install_python_agent
-  python_pip 'newrelic' do
+  python_package 'newrelic' do
     virtualenv new_resource.virtualenv if new_resource.virtualenv
     version new_resource.version if new_resource.version
+    python new_resource.python if new_resource.python
     action :install
   end
 end