+++ test/fixtures/cookbooks/newrelic_lwrp_test/recipes/agent_python_recipe.rb
@@ -11,5 +11,12 @@ node.default['newrelic']['application_monitoring']['error_collector']['enable'] 
 node.default['newrelic']['application_monitoring']['transaction_tracer']['enable'] = true
 node.default['newrelic']['application_monitoring']['enabled'] = 'true'
 
-include_recipe 'python'
+python_runtime '2' do
+  provider :system
+  setuptools_version true
+  virtualenv_version false
+  pip_version true
+  action :install
+end
+
 include_recipe 'newrelic::python_agent'